Revert "Revert "'add-Middleware'""
