+++ db.json
@@ -42,6 +42,16 @@
       "name": "si",
       "phone": "123",
       "id": "XUWh0UJRa"
+    },
+    {
+      "name": "Cuong",
+      "phone": "9977",
+      "id": "nHqLXtVH3"
+    },
+    {
+      "name": "Ninh",
+      "phone": "1231",
+      "id": "9k4XLyvzu"
     }
   ]
 }
