+++ routes/user.route.js
@@ -6,6 +6,7 @@ var express = require('express');
 var router = express.Router();
 
 var controller = require('../controller/user.controller');
+var validate = require('../validate/user.validate')
 
 
 
@@ -13,5 +14,5 @@ router.get('/', controller.index )
 router.get('/search', controller.search )
 router.get('/create', controller.create);
 router.get('/:id', controller.get)
-router.post('/create', controller.postCreate)
+router.post('/create', validate.postCreate ,controller.postCreate)
 module.exports = router;
